+++ lib/main.dart
@@ -1,13 +1,15 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_pokemon_app/pages/home_page.dart';
 import 'package:flutter_pokemon_app/pages/landing_page.dart';
+import 'package:flutter_pokemon_app/states/pokemon_cubit.dart';
 import 'package:flutter_pokemon_app/states/pokemon_list_cubit.dart';
 
 void main() {
-  runApp(
-    BlocProvider(create: (context) => PokemonListCubit(), child: const MyApp()),
-  );
+  runApp(const MyApp());
+  runApp(MultiBlocProvider(providers: [
+    BlocProvider(create: (context) => PokemonCubit()),
+    BlocProvider(create: (context) => PokemonListCubit()),
+  ], child: const MyApp()));
 }
 
 class MyApp extends StatelessWidget {
@@ -21,7 +23,7 @@ class MyApp extends StatelessWidget {
       theme: ThemeData(
         primarySwatch: Colors.red,
       ),
-      home: const LandingPage(),
+      home: LandingPage(),
     );
   }
 }
